(PUP-6220) Move Loader::TypedName out from Loader class
Before this commit, the full name of the TypedName was:

    Puppet::Pops::Loader::Loader::TypedName

This commit changes the name to:

    Puppet::Pops::Loader::TypedName

and moves the class to a separate file.
