+++ lib/puppet/pops/loader/puppet_function_instantiator.rb
@@ -60,13 +60,13 @@ class PuppetFunctionInstantiator
   end
 
   # Creates Function class and instantiates it based on a FunctionDefinition model
-  # @return [Array<Loader::TypedName, Functions.Function>] - array of 
+  # @return [Array<TypedName, Functions.Function>] - array of
   #   typed name, and an instantiated function with global scope closure associated with the given loader
   #
   def self.create_from_model(function_definition, loader)
     closure_scope = Puppet.lookup(:global_scope) { {} }
     created = create_function_class(function_definition, closure_scope)
-    typed_name = Loader::TypedName.new(:function, function_definition.name)
+    typed_name = TypedName.new(:function, function_definition.name)
     [typed_name, created.new(closure_scope, loader)]
   end
