+++ spec/unit/pops/loaders/static_loader_spec.rb
@@ -121,7 +121,7 @@ describe 'the static loader' do
   end
 
   def typed_name(type, name)
-    Puppet::Pops::Loader::Loader::TypedName.new(type, name)
+    Puppet::Pops::Loader::TypedName.new(type, name)
   end
 
   def resource_type(name)
